quick-note/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
pypdf
python-docx
chromadb
python-multipart
pyjwt
bcrypt
requests
